[frontend] change checklist page url pathname
