+++ frontend/public/locales/fr/checklist.json
